Allow for the creation of static views by setting `materialized = True`

+++ dataflow/operators/db_view.py
@@ -1,11 +1,12 @@
 create_view = """
-    DROP VIEW IF EXISTS
-        {{ view_name }}_{{ (
-            macros.datetime.strptime(ds, '%Y-%m-%d') +
-            macros.dateutil.relativedelta.relativedelta(months=+1, days=-1)
-        ).date() | replace('-', '_') }};
-
-    CREATE VIEW
+    {% if not materialized %}
+        DROP VIEW IF EXISTS
+            {{ view_name }}_{{ (
+                macros.datetime.strptime(ds, '%Y-%m-%d') +
+                macros.dateutil.relativedelta.relativedelta(months=+1, days=-1)
+            ).date() | replace('-', '_') }};
+    {% endif %}
+    CREATE{% if materialized %} MATERIALIZED{%endif %} VIEW
         {{ view_name }}_{{ (
             macros.datetime.strptime(ds, '%Y-%m-%d') +
             macros.dateutil.relativedelta.relativedelta(months=+1, days=-1)
